unint8 variable

+++ Intro/Variables/main.go
@@ -3,6 +3,6 @@ package main
 import "fmt"
 
 func main() {
-	nine := 9
+	nine := uint8(9)
 	fmt.Printf("Value: %d\n", nine)
 }
